backend/unsplash/requirements.txt
djangorestframework
django-cors-headers
psycopg2
djangorestframework-simplejwt